--- src/CreateSurvey/php/CreateSurveyAction.php
@@ -1,5 +1,0 @@
-<?php
-//actions when form submitted
-//Not yet implemented
-
-?>
